// src/api/generated/headscale.ts
export interface components {
    schemas: {
// ...
        v1PreAuthKey: {
            user?: components["schemas"]["v1User"];
            /** Format: uint64 */
            id?: string;
            key?: string;
            reusable?: boolean;
            ephemeral?: boolean;
            used?: boolean;
            /** Format: date-time */
            expiration?: string;
            /** Format: date-time */
            createdAt?: string;
            aclTags?: string[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}